Add license-eye checks to our CI. (#59)
* Add license-eye checks to our CI.

It's nice to run it on the local box, but it needs to be automated so
that we can enforce things.

* Fix licenses.

* Hopefully clean up some false positives.

I don't know that we **have** to put the license headers on the
package.json; I've never seen it.

* Add header to npmrc.

* Add header to schema.prisma.

* ignoring node modules from licence-eye

* header template.

I thought they were suppoed to manage that.

* Remove template.

That is apparently supposed to come from them, so it's their problem to fix.

* exclude schema.prisma from license-eye

* exclude schema.prisma from license-eye

* Let's try the main branch of the plugin.

This is what I am using right now, so that may fix the template
problem. Other than that, we might need to wait for a fix.

* Fix license heaers.

* Put the last of the ignore into place.

This should fix up the bulk of the errors.

* Fix some fo the config warnings, hopefully.

I don't see why these would be necessary, but let's add the to be sure
so as to squelch those warnings.

* Let's actually check and not fix.

I need more caffein!

* Similar fixes for the dependency checking.

* VIgnore the vercel file.

---------

Co-authored-by: DinoZoidDev <38674879+DinoZoidDev@users.noreply.github.com>
Co-authored-by: dinozoiddev <realdinozoid@gmail.com>